refactor: move divina cover reader to native containers
Replace the SwiftUI-driven DIVINA cover path with native UIKit and AppKit cover containers, shared paged page content views, and native slot management.

Fix stale loading state, reduce redundant slot refreshes, and correct first-transition animation issues on iOS and macOS.
